explicit type conversion

+++ src/main.rs
@@ -1,8 +1,8 @@
 fn main() {
-    let x = 255u8; 
+    let x = 255u32; 
     let y = 10_u8;  
 
-    let z = x % y;
+    let z = x % y as u32;
     println!("Hello, world! x:{} y:{} z:{}", x, y, z);
